fix: remove custom singal handling, use context and waitgroup, use http server for graceful shutdown
Removed the custom singal handling and pass a context instead.
the singals are handled in main and the context is canceled as normal

use gin router as handler to http server so we can use the graceful
shutdown functionality. This allows us to be a bit nicer to open
connections when the context is canceled